+++ Jubging/app/src/main/res/layout/activity_main.xml
@@ -47,6 +47,8 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent" />
 
+
+
     </androidx.constraintlayout.widget.ConstraintLayout>
